extract TableDecorator from ColoredSlimTable
